don't redownload

+++ scripts/install.cncfm.debian.sh
@@ -49,7 +49,10 @@ mkdir -p $APPDIR/logs
 mkdir -p $APPDIR/USERS/default
 mkdir -p $APPDIR/bin
 cp -rf $MEDIR/../* $APPDIR
-wget "https://inkscape.org/gallery/item/37359/Inkscape-b0a8486-x86_64.AppImage" -O $APPDIR/bin/Inkscape.AppImage
+if [ ! -f $APPDIR/bin/Inkscape.AppImage ]
+then
+    wget "https://inkscape.org/gallery/item/37359/Inkscape-b0a8486-x86_64.AppImage" -O $APPDIR/bin/Inkscape.AppImage
+fi
 
 
 # Ask user which example config to start
